src/requirements.txt
pillow
pygame
PyOpenGL
